@@ -1,14 +1,11 @@
-﻿using System;
+﻿namespace App;
 
-namespace App
+public static class Extensions
 {
-    public static class Extensions
+    public static void WriteLine(this ConsoleColor color, object value)
     {
-        public static void WriteLine(this ConsoleColor color, object value)
-        {
-            Console.ForegroundColor = color;
-            Console.WriteLine(value);
-            Console.ResetColor();
-        }
+        Console.ForegroundColor = color;
+        Console.WriteLine(value);
+        Console.ResetColor();
     }
 }
